merge da super integração (falta testes)

See merge request sistema-de_gerenciador_bicicletario/ms-aluguel!20

+++ src/main/java/com/sistema_bicicletario/ms_aluguel/dtos/AluguelDTO.java
@@ -11,7 +11,7 @@ public class AluguelDTO {
     private LocalDateTime horaInicio;
     private LocalDateTime horaFim;
     private Integer trancaFim;
-    private Integer cobranca;
+    private Long cobranca;
     private Integer ciclista;
     private Integer trancaInicio;
 }
